rmv menu.rb
